:recycle: detach Slider file.
detach file by slider component.
