+++ src/components/common/Slider/SlideList.tsx
@@ -1,0 +1,23 @@
+import { css } from 'styled-components';
+
+import { useSliderInfoContext } from '@components/common/Slider/context/SliderInfoContext';
+
+const SlideList = ({ children, ...restProps }) => {
+  const {
+    options: { slidesMargin },
+  } = useSliderInfoContext();
+
+  return (
+    <ul
+      css={css`
+        display: flex;
+        gap: ${slidesMargin};
+      `}
+      {...restProps}
+    >
+      {children}
+    </ul>
+  );
+};
+
+export default SlideList;
