+++ src/components/common/Slider/index.tsx
@@ -1,120 +1,19 @@
-import {
-  createContext,
-  useContext,
-  Dispatch,
-  useReducer,
-  ReactNode,
-  Children,
-} from 'react';
+import { ReactNode } from 'react';
 import { css } from 'styled-components';
 
-import throttle from '@utils/eventDelay';
-
-type ChildrenType = ReactNode | ReactNode[];
-
-type SliderOptions = {
-  slidesToShow: number;
-  slidesToScroll: number;
-  slidesMargin: string;
-  arrows: boolean;
-  speed: number;
-  initialSlide: number;
-};
-
-type SliderInfoContextValues = {
-  options: SliderOptions;
-  slidesLength: number;
-};
-
-const SliderInfoContext = createContext<SliderInfoContextValues | null>(null);
-
-const useSliderInfoContext = () => {
-  const context = useContext(SliderInfoContext);
-
-  if (!context)
-    throw new Error(
-      'useSliderInfoContext should be used within SliderInfoContext.Provider',
-    );
-
-  return context;
-};
-
-type SliderItemIndexContextValues = {
-  currentIndex: number;
-};
-
-type SliderItemIndexAction = {
-  type: 'INCREASE_CURRENT_INDEX' | 'DECREASE_CURRENT_INDEX';
-  slidesToScroll: number;
-};
-
-const reducer = (
-  { currentIndex }: SliderItemIndexContextValues,
-  action: SliderItemIndexAction,
-) => {
-  const { type, slidesToScroll } = action;
-
-  switch (type) {
-    case 'INCREASE_CURRENT_INDEX':
-      return { currentIndex: currentIndex + slidesToScroll };
-    case 'DECREASE_CURRENT_INDEX':
-      return { currentIndex: currentIndex - slidesToScroll };
-    default:
-      throw new Error(`There is no type '${type}'. Please check the type.`);
-  }
-};
-
-const SliderItemIndexContext =
-  createContext<
-    [SliderItemIndexContextValues, Dispatch<SliderItemIndexAction>] | null
-  >(null);
-
-const useSliderItemIndexContext = () => {
-  const context = useContext(SliderItemIndexContext);
-
-  if (!context)
-    throw new Error(
-      'useSliderItemIndexContext should be used within SliderItemIndexContext.Provider',
-    );
-
-  return context;
-};
-
-type SliderProviderProps = {
-  options: SliderOptions;
-  slidesLength: number;
-  children: ChildrenType;
-};
-
-const SliderProvider = ({
-  options,
-  slidesLength,
-  children,
-}: SliderProviderProps) => {
-  const sliderInfo = {
-    options,
-    slidesLength,
-  };
-
-  const sliderItemIndexReducer = useReducer(reducer, {
-    currentIndex: options.initialSlide,
-  });
-
-  return (
-    <SliderInfoContext.Provider value={sliderInfo}>
-      <SliderItemIndexContext.Provider value={sliderItemIndexReducer}>
-        {children}
-      </SliderItemIndexContext.Provider>
-    </SliderInfoContext.Provider>
-  );
-};
+import { SliderOptions } from '@components/common/Slider/context/SliderInfoContext';
+import SliderProvider from '@components/common/Slider/context/SliderProvider';
+import NextButton from '@components/common/Slider/NextButton';
+import PrevButton from '@components/common/Slider/PrevButton';
+import Slide from '@components/common/Slider/Slide';
+import SlideList from '@components/common/Slider/SlideList';
 
 type SliderProps = {
   options?: Partial<SliderOptions>;
-  children: ChildrenType;
+  children: ReactNode | ReactNode[];
 };
 
-const defaultSliderOptions = {
+const defaultOptions = {
   slidesToShow: 1,
   slidesToScroll: 1,
   slidesMargin: '0px',
@@ -123,25 +22,26 @@ const defaultSliderOptions = {
   initialSlide: 0,
 };
 
-const Slider = ({ options = defaultSliderOptions, children }: SliderProps) => {
-  const sliderOptions = { ...defaultSliderOptions, ...options };
+const Slider = ({ options = defaultOptions, children }: SliderProps) => {
+  const sliderOptions = { ...defaultOptions, ...options };
 
-  const getSlidesLength = () => {
-    const sliderList = Children.toArray(children).find((child: ReactNode) => {
+  const getSlideLength = () => {
+    const Children = Array.isArray(children) ? children : [children];
+    const slideList = Children.find((child) => {
       if (
-        child?.type.name === 'SliderList' ||
-        child?.type.target.name === 'SliderList'
+        child?.type.name === 'SlideList' ||
+        child?.type.target.name === 'SlideList'
       )
         return true;
       return false;
     });
-    const slidesLength = sliderList ? sliderList.props.children.length : 0;
+    const SlideLength = slideList ? slideList.props.children.length : 0;
 
-    return slidesLength;
+    return SlideLength;
   };
 
   return (
-    <SliderProvider options={sliderOptions} slidesLength={getSlidesLength()}>
+    <SliderProvider options={sliderOptions} SlideLength={getSlideLength()}>
       <div
         css={css`
           position: relative;
@@ -155,131 +55,9 @@ const Slider = ({ options = defaultSliderOptions, children }: SliderProps) => {
   );
 };
 
-const SliderList = ({ children, ...restProps }) => {
-  const {
-    options: { slidesMargin },
-  } = useSliderInfoContext();
-
-  return (
-    <ul
-      css={css`
-        display: flex;
-        gap: ${slidesMargin};
-      `}
-      {...restProps}
-    >
-      {children}
-    </ul>
-  );
-};
-
-const SliderItem = ({ children, ...restProps }) => {
-  const [{ currentIndex }] = useSliderItemIndexContext();
-  const {
-    options: { slidesToShow, slidesMargin, speed },
-  } = useSliderInfoContext();
-
-  return (
-    <li
-      css={css`
-        flex: 0 0 auto;
-        width: calc(
-          100% / ${slidesToShow} -
-            (${slidesMargin} * (${slidesToShow - 1}) / ${slidesToShow})
-        );
-        transform: translate(
-          calc((100% + ${slidesMargin}) * ${-currentIndex}),
-          0
-        );
-        transition: transform calc(${speed}s / 1000) ease-in-out;
-      `}
-      {...restProps}
-    >
-      {children}
-    </li>
-  );
-};
-
-const SliderPrevButton = ({ children, ...restProps }) => {
-  const [{ currentIndex }, dispatch] = useSliderItemIndexContext();
-  const {
-    options: { slidesToScroll, arrows, speed },
-  } = useSliderInfoContext();
-  const limitIndex = 0;
-  const disabled = currentIndex <= limitIndex;
-
-  const handlePrevButtonClick = () => {
-    if (disabled) return;
-
-    const decreasedIndex =
-      currentIndex - slidesToScroll > limitIndex
-        ? slidesToScroll
-        : currentIndex - limitIndex;
-
-    throttle(
-      () =>
-        dispatch({
-          type: 'DECREASE_CURRENT_INDEX',
-          slidesToScroll: decreasedIndex,
-        }),
-      speed,
-    );
-  };
-
-  return arrows ? (
-    <button
-      type="button"
-      onClick={handlePrevButtonClick}
-      disabled={disabled}
-      {...restProps}
-    >
-      {children}
-    </button>
-  ) : null;
-};
-
-const SliderNextButton = ({ children, ...restProps }) => {
-  const [{ currentIndex }, dispatch] = useSliderItemIndexContext();
-  const {
-    options: { slidesToShow, slidesToScroll, arrows, speed },
-    slidesLength,
-  } = useSliderInfoContext();
-  const limitIndex = slidesLength - slidesToShow;
-  const disabled = currentIndex >= limitIndex;
-
-  const handleNextButtonClick = () => {
-    if (disabled) return;
-
-    const increasedIndex =
-      currentIndex + slidesToScroll < limitIndex
-        ? slidesToScroll
-        : limitIndex - currentIndex;
-
-    throttle(
-      () =>
-        dispatch({
-          type: 'INCREASE_CURRENT_INDEX',
-          slidesToScroll: increasedIndex,
-        }),
-      speed,
-    );
-  };
-
-  return arrows ? (
-    <button
-      type="button"
-      onClick={handleNextButtonClick}
-      disabled={disabled}
-      {...restProps}
-    >
-      {children}
-    </button>
-  ) : null;
-};
-
-Slider.List = SliderList;
-Slider.Item = SliderItem;
-Slider.PrevButton = SliderPrevButton;
-Slider.NextButton = SliderNextButton;
+Slider.List = SlideList;
+Slider.Item = Slide;
+Slider.PrevButton = PrevButton;
+Slider.NextButton = NextButton;
 
 export default Slider;
